Inclusion de toutes les méthodes de dessins dans SWIG
Toutes, non, mire et dessine_liste sont soit liés à la physique du robot
(taille des bras), soit à des données externes, e.g. liste.h
